Select/configure ondemand cpu governor
Select/configure the ondemand cpu governor with a combination of
udev/tmpfiles rules.

The selection/configuration of the ondemand cpu governor was previously
achieved with the raspi-config SysVinit script (see raspi-config
package). /etc/default/cpu_governor is no longer supported as a
configuration mechanism;
/etc/tmpfiles.d/raspberrypi-sys-mods-ondemand-governor.conf should be
used to override the configuration provided by this package.